Extract type to its own file
